ChatBot/requirements.txt
# ==== FastAPI & Server ====
fastapi==0.111.0
uvicorn==0.30.1
pyngrok==5.3.0
pydantic==2.7.1
requests==2.31.0

# ==== CrewAI Core ====
crewai==0.28.8
langchain==0.2.1
langchain-groq==0.1.3
langchain-community==0.2.1

# ==== RAG System ====
faiss-cpu==1.8.0
langchainhub==0.1.15
sentence-transformers==2.6.1
transformers==4.41.2
torch==2.3.0
accelerate==0.30.1

# ==== Groq API ====
groq==0.5.0

# ==== Utilities ====
python-dotenv==1.0.1
typing-extensions==4.11.0
